+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   - updated language endpoint changed structure
 - Service Release Process
   - Skip technical integration for serviceTypeIds(consultancy)
+- SSI Certificat Credential Adminboard
+  - Add new page and created UI
+  - Add Backend Logic to fetch certificates
+  - Add Confirm/Decline business logic
 
 ### Feature